Simple Receipt Generator Added

+++ Item.cs
@@ -4,10 +4,10 @@ public class Item
 {
     public string ItemName { get; }
     public string ItemDescription { get; }
-    public double ItemPrice { get; }
+    public decimal ItemPrice { get; }
     public int ItemIndex { get; }
 
-    public Item(string itemName, string itemDeskription, double itemPrice, int itemIndex) //Item <Class> constructor.
+    public Item(string itemName, string itemDeskription, decimal itemPrice, int itemIndex) //Item <Class> constructor.
     {
         ItemName = itemName;
         ItemDescription = itemDeskription;
